# [ThreadPool] Reduce lock contention