--- a/sources/core/Xenko.Core/Threading/ThreadPool.cs
+++ b/sources/core/Xenko.Core/Threading/ThreadPool.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Xenko contributors (https://xenko.com) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
 using System;
-using System.Collections.Generic;
+using System.Collections.Concurrent;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
@@ -25,10 +25,9 @@ namespace Xenko.Core.Threading
         private readonly Action<object> cachedTaskLoop;
 
         private readonly int maxThreadCount = Environment.ProcessorCount + 2;
-        private readonly Queue<Action> workItems = new Queue<Action>();
-        private readonly ManualResetEvent workAvailable = new ManualResetEvent(false);
+        private readonly ConcurrentBag<Action> workItems = new ConcurrentBag<Action>();
+        private readonly SemaphoreSlim workAvailable = new SemaphoreSlim(0, int.MaxValue);
 
-        private SpinLock spinLock = new SpinLock();
         private int busyCount;
         private int aliveCount;
 
@@ -40,34 +39,23 @@ namespace Xenko.Core.Threading
 
         public void QueueWorkItem([NotNull] [Pooled] Action workItem)
         {
-            bool lockTaken = false;
             bool startNewTask = false;
             PooledDelegateHelper.AddReference(workItem);
-            try
-            {
-                spinLock.Enter(ref lockTaken);
-                workItems.Enqueue(workItem);
-                workAvailable.Set();
 
-                // We're only locking when potentially increasing aliveCount as we
-                // don't want to go above our maximum amount of threads.
-                int curBusyCount = Interlocked.CompareExchange(ref busyCount, 0, 0);
-                int curAliveCount = Interlocked.CompareExchange(ref aliveCount, 0, 0);
-                if (curBusyCount + 1 >= curAliveCount && curAliveCount < maxThreadCount)
-                {
-                    // Start threads as busy otherwise only one thread will be created 
-                    // when calling this function multiple times in a row
-                    Interlocked.Increment(ref busyCount);
-                    Interlocked.Increment(ref aliveCount);
-                    startNewTask = true;
-                }
-            }
-            finally
+            workItems.Add(workItem);
+            workAvailable.Release(1);
+
+            // We're only locking when potentially increasing aliveCount as we
+            // don't want to go above our maximum amount of threads.
+            int curBusyCount = Interlocked.CompareExchange(ref busyCount, 0, 0);
+            int curAliveCount = Interlocked.CompareExchange(ref aliveCount, 0, 0);
+            if (curBusyCount + 1 >= curAliveCount && curAliveCount < maxThreadCount)
             {
-                if (lockTaken)
-                {
-                    spinLock.Exit(true);
-                }
+                // Start threads as busy otherwise only one thread will be created 
+                // when calling this function multiple times in a row
+                Interlocked.Increment(ref busyCount);
+                Interlocked.Increment(ref aliveCount);
+                startNewTask = true;
             }
             // No point in wasting spins on the lock while creating the task
             if (startNewTask)
@@ -84,39 +72,7 @@ namespace Xenko.Core.Threading
                 long lastWorkTS = Stopwatch.GetTimestamp();
                 while (true)
                 {
-                    Action workItem = null;
-                    bool lockTaken = false;
-                    try
-                    {
-                        spinLock.Enter(ref lockTaken);
-                        if (workItems.Count > 0)
-                        {
-                            workItem = workItems.Dequeue();
-                            if (workItems.Count == 0)
-                            {
-                                workAvailable.Reset();
-                            }
-                        }
-                    }
-                    finally
-                    {
-                        if (lockTaken)
-                        {
-                            spinLock.Exit(true);
-                        }
-                    }
-                    
-                    if (workItem == null)
-                    {
-                        bool idleForTooLong = Stopwatch.GetTimestamp() - lastWorkTS > MaxIdleTimeTS;
-                        // Wait for another work item to be (potentially) available
-                        if (idleForTooLong || workAvailable.WaitOne(MaxIdleTimeInMS) == false)
-                        {
-                            // No work given in the last MaxIdleTimeTS, close this task
-                            return;
-                        }
-                    }
-                    else
+                    if (workItems.TryTake( out Action workItem ))
                     {
                         Interlocked.Increment(ref busyCount);
                         try
@@ -131,6 +87,16 @@ namespace Xenko.Core.Threading
                         }
                         PooledDelegateHelper.Release(workItem);
                         lastWorkTS = Stopwatch.GetTimestamp();
+                    }
+                    else
+                    {
+                        bool idleForTooLong = Stopwatch.GetTimestamp() - lastWorkTS > MaxIdleTimeTS;
+                        // Wait for another work item to be (potentially) available
+                        if (idleForTooLong || workAvailable.Wait(MaxIdleTimeInMS) == false)
+                        {
+                            // No work given in the last MaxIdleTimeTS, close this task
+                            return;
+                        }
                     }
                 }
             }