# PR Details
Replaced locking with semaphore+bag. Only a single thread is woken up per job queued up, they don't all hammer on the lock blocking further work from being queued up, this version should be more efficient and faster even more so for machines with many cores.

## Description
See above.

## Related Issue
None.

## Motivation and Context
Performances.

## Types of changes
- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist
- [ ] My change requires a change to the documentation.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.