@phr00t 2019-10-23
In my testing, this ThreadPool system was significantly slower than what I have implemented. I get ~80 FPS on my benchmark application with my current ThreadPool (on my fork). Using this one, I got about ~73-74 FPS. What was worse, is when nearly nothing was rendering, slowdowns were even more apparent and I couldn't get above 120 FPS. My current ThreadPool would max at my display refresh rate of 144 FPS in the same situation. My guess is, ConcurrentBag isn't really that fast and may do some internal locking itself.

Tested with an i7-9750H with the "DemoArmyBuilder" application.

We may have already squeezed out as much performance from this as possible... :-/