@phr00t 2019-10-23
I'm using a different ThreadPool than the one you are using to compare with:

https://github.com/phr00t/FocusEngine/blob/master/sources/core/Xenko.Core/Threading/ThreadPool.cs

... you might want to test with mine above and see how it compares to the ones you are using.