@Eideren 2019-10-23
@phr00t Yours is pretty much the same thing as what's on ``master``, tested it anyway just in case and I reach around 3.4 ms / frame (~290 fps) with multiple spikes per second of 20 ms/frame and above, same as on master.
While this one is 2.2 ms / frame (~440 fps) without any significant spikes (!).
I swapped the code around a bit and ConcurrentBag doesn't do much anymore so I implemented the queue back, try that one out as well if you can.