@phr00t 2019-10-23
Ran more tests with DemoArmyBuilder, Vulkan, using your latest ThreadPool. Here are frame rates calculated every second:
75.48298
76.238
69.67778
75.63609
73.79427
76.46061
75.23015
76.45001
76.34172
75.43897

Here is using my ThreadPool:
79.37527
79.86379
79.12566
78.91089
78.65545
80.07073
78.21236
79.91784
79.42142
79.67166
78.60777

My guess is, you are testing a very simple scene with little work, hence getting ~300-400 FPS. I'm testing a very heavy scene. I'm also testing with Vulkan, where you may be testing with DirectX. Your system may work faster on light loads, but mine works better on heavy loads. This is all a guess, though... not sure what is going on, but your method doesn't seem overall faster for me.