@Eideren 2019-10-23
@phr00t Is that your 'DemoArmyBuilder' ? https://github.com/romanstetsenko/DemoArmyBuilder doesn't seem any heavier than what I'm running, I hover around 220 minimum.