@phr00t 2019-10-23
What are your system specs? Do you zoom out to view all of the thousands of objects generated (which is where I measure FPS). I did add some post processing to mine too, but don't understand how you are getting 220 FPS on it. I am running the program in debug due to printing FPS to console (since printing debug text in Vulkan doesn't work). Engine is built in release mode.