@Eideren 2019-10-24
@phr00t 980ti + ryzen 2700x
![image](https://user-images.githubusercontent.com/5742236/67442452-2359d200-f601-11e9-8918-162abff1c0e3.png)
Program built in release, the only changes I made where to set the bool 'enable' on the gameprofiler to ``true`` and set the packages to 3.1.0.1-beta02-0752 (doesn't contain this PR's changes).
I'm not using my profiling tools so I can't say if those are accurate but the maximum delta of that framerate is at most +-50 fps from what's shown on the screen.