@phr00t 2019-10-24
Your CPU is faster than mine: https://www.cpubenchmark.net/compare/AMD-Ryzen-7-2700X-vs-Intel-i7-9750H/3238vs3425

Although my RTX 2060 edges out your GPU by a little...

Neither of these explain why you get a whopping 730(!) FPS with that test scene, though...

My FPS counter is calculated like this:

`        public float accum;
        public int frames;

        public override void Update()
        {
            accum += (float)Game.DrawTime.Elapsed.TotalSeconds;
            frames++;
            if (frames > 60) {
                System.Diagnostics.Debug.WriteLine(1f / (accum / frames));
                accum = 0f;
                frames = 0;
            }`

... so averages 60 frames worth of DrawTime.Elapsed.TotalSeconds. Using UpdateTime gives me the same results:

![Untitled](https://user-images.githubusercontent.com/5983470/67446661-525f4c00-f5df-11e9-826a-f60b98c158db.png)