@phr00t 2019-10-24
Just to confirm, your scene isn't instancing all of those objects, right? Mine are all separate objects to maximize overhead for benchmarking.