@Eideren 2019-10-24
@phr00t I don't have a clue, as I said, I just pulled the repo and the only changes I made where to set the packages to 3.1.0.1-beta02-0752, if you have a specific setup, please go ahead and send it to me.
Your logic does output framerates above 650 per seconds (once I removed vsync/pulled xenko's profiler(``LCtrl``+``LShift``+``P``)), something might be wrong in the way you're profiling or in your branch if your ``DemoArmyBuilder`` matches the repo I linked above.