@phr00t 2019-10-24
I fixed the Xenko profiler with Vulkan on my fork, but I get the same results from it (~79 FPS) on Release.

Roman initially created the DemoArmyBuilder to be very demanding and slow, and he took a screenshot of it running on master @ 48 FPS:

https://github.com/romanstetsenko/DemoArmyBuilder/blob/master/Why.png

... which is why my numbers improving upon those, but still in the same ballpark, seemed to make more sense. 650-700+ FPS with 5000 seperate objects on screen doesn't sound right to me, but I don't know what is going on.