@SebastianStehle 2019-10-24
I would also consider to pass in an optional state object.

This can reduce the allocations that are needed to capture the variables. Not sure if it has an impact in this case.