@Eideren 2019-10-24
@phr00t Well, tbf 5000 at max 80 fps isn't very impressive if all they are doing is rendering, even more so on Vulkan/DX12. Did you try to pull it again and profile it with xenko's master and yours ? And send me your version please, I can't really move forward without putting this issue to rest.