@phr00t 2019-10-24
I can't build master after a fresh clone, I get junk like this:

1>This version of D:\Phr00tsSoftware\xenko-official\deps\LLVM\clang.exe is not compatible with the version of Windows you're running. Check your computer's system information and then contact the software publisher.
1>D:\Phr00tsSoftware\xenko-official\sources\native\Xenko.Native.targets(125,5): error MSB3073: The command ""D:\Phr00tsSoftware\xenko-official\deps\\LLVM\clang.exe" -gcodeview -fno-ms-extensions -nobuiltininc -nostdinc++  -Wno-ignored-attributes -Wno-delete-non-virtual-dtor -Wno-macro-redefined -I"D:\Phr00tsSoftware\xenko-official\deps\\NativePath" -I"D:\Phr00tsSoftware\xenko-official\deps\\NativePath\standard" -O0 -g -DNEED_DLL_EXPORT -o "D:\Phr00tsSoftware\xenko-official\sources\core\Xenko.Core\obj\Debug\netstandard2.0\win\\x86\CoreNative.obj" -c "D:\Phr00tsSoftware\xenko-official\sources\core\Xenko.Core\Native\CoreNative.c" -fms-extensions -DWINDOWS_DESKTOP -target i686-pc-windows-msvc" exited with code 216.
1>Done building project "Xenko.Core.csproj" -- FAILED.

Here is my DemoArmyBuilder: https://drive.google.com/open?id=1DRf4PhOrhG3wSMuxWHFLGcwKLwB8B4hO

You will probably need to remove the OverrideDefaultSettings call in the startup .cs, which sets resolution on my fork. I also have some post processing effects enabled (fog and outlining), but those will probably be just skipped when loading it on yours.

Perhaps try not using Xenko's profiler to track frame rate, but have a little accumulator like I did that prints out to a debug console? I remember you did a bunch of GameBase refactoring that might have altered the profiler (although that might be another branch, so it wouldn't apply).

Nobody said 80 FPS is impressive, but it is more consistent with the creator's benchmark results than 700+ FPS. I remember getting below 48 at first, but finding improvements with Vulkan and ThreadPool to get it up to 80 FPS.