@Eideren 2019-10-26
@phr00t Welp, didn't really see them in the sea of sub-directories that that dir contains. I get around 64, which is within the margin of error to be on par with master's dx11.

So I made a benchmark and improved the implementation, this is ready for review/merge.

Here's how this new one fares against the old one and dotnet's threadpool.
```
|                Method | JobsDuration |         Mean |        Error |       StdDev | Rank |
|---------------------- |------------- |-------------:|-------------:|-------------:|-----:|
|               NewPush |           10 |     563.4 us |     9.404 us |     8.796 us |    1 |
|            DotnetPush |           10 |   1,019.9 us |     7.681 us |     6.809 us |    4 |
|               OldPush |           10 |     937.5 us |    17.985 us |    21.410 us |    2 |
-
|             NewSpaced |           10 |   1,393.9 us |    29.361 us |    46.569 us |    5 |
|          DotnetSpaced |           10 |   1,442.8 us |     9.318 us |     8.716 us |    6 |
|             OldSpaced |           10 |   1,947.5 us |    36.147 us |    35.501 us |    8 |
-
|    NewDispatchAndWait |           10 |   9,014.4 us |    30.828 us |    28.836 us |   10 |
| DotnetDispatchAndWait |           10 |  15,553.5 us |   166.248 us |   147.375 us |   13 |
|    OldDispatchAndWait |           10 |  17,082.7 us |   269.837 us |   225.326 us |   14 |
-
|               NewPush |          100 |     979.1 us |     8.832 us |     8.261 us |    3 |
|            DotnetPush |          100 |   1,011.2 us |    19.415 us |    19.938 us |    4 |
|               OldPush |          100 |   1,647.9 us |    31.267 us |    32.109 us |    7 |
-
|             NewSpaced |          100 |   8,904.7 us |   144.175 us |   127.808 us |   10 |
|          DotnetSpaced |          100 |  10,939.6 us |   170.793 us |   159.759 us |   11 |
|             OldSpaced |          100 |  12,086.7 us |   239.107 us |   558.904 us |   12 |
-
|    NewDispatchAndWait |          100 |  20,218.5 us |    62.375 us |    48.698 us |   15 |
| DotnetDispatchAndWait |          100 |  23,834.8 us |    36.913 us |    30.824 us |   16 |
|    OldDispatchAndWait |          100 |  49,436.9 us |   971.530 us |   908.770 us |   17 |
-
|               NewPush |         1000 |   8,429.8 us |   165.064 us |   293.400 us |    9 |
|            DotnetPush |         1000 |   8,611.5 us |   108.883 us |    96.522 us |    9 |
|               OldPush |         1000 |   9,025.7 us |    72.784 us |    68.082 us |   10 |
-
|             NewSpaced |         1000 |  81,109.7 us |   325.034 us |   288.134 us |   18 |
|          DotnetSpaced |         1000 |  86,667.2 us | 1,835.622 us | 2,113.905 us |   19 |
|             OldSpaced |         1000 |  85,738.8 us |   140.944 us |   124.944 us |   19 |
-
|    NewDispatchAndWait |         1000 | 137,554.5 us |   788.102 us |   698.632 us |   20 |
| DotnetDispatchAndWait |         1000 | 138,408.3 us | 1,066.573 us |   945.489 us |   20 |
|    OldDispatchAndWait |         1000 | 280,874.3 us | 2,762.420 us | 2,583.969 us |   21 |
```
- Dotnet -> ``System.Threading.ThreadPool``
- Old -> What's in master right now
- New -> This PR's changes

- Push: Enqueues 1000 jobs in a tight loop, wait for all of them to finish.
  - Tests heavy locking contentions.
- Spaced: Enqueues 10 jobs, execute one job, repeat. After 1000, wait for all of them to finish.
  - Tests lesser locking amount. (not really suited to machines with less than 10 cores/hw threads)
- DispatchAndWait: Enqueues ``Environment.ProcessorCount-1`` jobs, wait for those to finish, repeat 1000 times.
  - Test threads waking up and response speed.
- JobsDuration: Duration in amount of operations for each job enqueued.

As with all multithreading benchmark on windows, don't take those numbers too seriously, it's faster in all cases but the ratio by which it is varies widely based on windows' mood at the time.
Here it is if you want to run it: [ConsoleApp1.zip](https://github.com/xenko3d/xenko/files/3774492/ConsoleApp1.zip). Do note that it takes an extensive amount of time (upwards of 10 minutes on my pc) and that you should, of course, close everything while it is running as we need all threads to be idle to have an accurate benchmark.

A couple of notes on things I tried which didn't end up improving performances:
- Other queue types:
   - ConcurrentBag: Slower than ``ConcurrentQueue`` for some reason.
   - Queue with SpinLock: Same as above, significantly slower.
 - SpinWait instead of similar mechanism in UnfairSemaphore:
   - Significantly slower, most likely because SpinWait will ``Thread.Sleep(1)`` every once in a while, which is pointless in our case.
- Weird legacy ``Environment.ProcessorCount + 2`` to just ``Environment.ProcessorCount``
   - Did not really have any significant changes to performance so I'll leave it as is as I don't really understand how it would help.